Move common PlatformInformation to core
