+++ adal/src/Microsoft.IdentityModel.Clients.ActiveDirectory/Platforms/net45/PlatformInformation.cs
@@ -44,89 +44,6 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory.Internal.Platform
             return "PCL.Desktop";
         }
 
-        public override async Task<string> GetUserPrincipalNameAsync()
-        {
-            return await Task.Factory.StartNew(() =>
-            {
-                const int NameUserPrincipal = 8;
-                uint userNameSize = 0;
-                WindowsNativeMethods.GetUserNameEx(NameUserPrincipal, null, ref userNameSize);
-                if (userNameSize == 0)
-                {
-                    throw new AdalException(AdalError.GetUserNameFailed, new Win32Exception(Marshal.GetLastWin32Error()));
-                }
-
-                StringBuilder sb = new StringBuilder((int) userNameSize);
-                if (!WindowsNativeMethods.GetUserNameEx(NameUserPrincipal, sb, ref userNameSize))
-                {
-                    throw new AdalException(AdalError.GetUserNameFailed, new Win32Exception(Marshal.GetLastWin32Error()));
-                }
-
-                return sb.ToString();
-            }).ConfigureAwait(false);
-        }
-
-        public override string GetEnvironmentVariable(string variable)
-        {
-            string value = Environment.GetEnvironmentVariable(variable);
-            return !string.IsNullOrWhiteSpace(value) ? value : null;
-        }
-
-        public override string GetProcessorArchitecture()
-        {
-            return WindowsNativeMethods.GetProcessorArchitecture();
-        }
-
-        public override string GetOperatingSystem()
-        {
-            return Environment.OSVersion.ToString();
-        }
-
-        public override string GetDeviceModel()
-        {
-            // Since ADAL .NET may be used on servers, for security reasons, we do not emit device type.
-            return null;
-        }
-
-        public override async Task<bool> IsUserLocalAsync(RequestContext requestContext)
-        {
-            return await Task.Factory.StartNew(() =>
-            {
-                WindowsIdentity current = WindowsIdentity.GetCurrent();
-                if (current != null)
-                {
-                    string prefix = WindowsIdentity.GetCurrent().Name.Split('\\')[0].ToUpperInvariant();
-                    return prefix.Equals(Environment.MachineName.ToUpperInvariant(), StringComparison.OrdinalIgnoreCase);
-                }
-
-                return false;
-            }).ConfigureAwait(false);
-        }
-
-        public override bool IsDomainJoined()
-        {
-            bool returnValue = false;
-            try
-            {
-                WindowsNativeMethods.NetJoinStatus status;
-                IntPtr pDomain;
-                int result = WindowsNativeMethods.NetGetJoinInformation(null, out pDomain, out status);
-                if (pDomain != IntPtr.Zero)
-                {
-                    WindowsNativeMethods.NetApiBufferFree(pDomain);
-                }
-
-                returnValue = result == WindowsNativeMethods.ErrorSuccess &&
-                              status == WindowsNativeMethods.NetJoinStatus.NetSetupDomainName;
-            }
-            catch (Exception)
-            {
-                // ignore the exception as the result is already set to false
-            }
-
-            return returnValue;
-        }
-
         public override void AddPromptBehaviorQueryParameter(IPlatformParameters parameters, DictionaryRequestParameters authorizationRequestParameters)
         {
             PlatformParameters authorizationParameters = (parameters as PlatformParameters);
