+++ adal/src/Microsoft.IdentityModel.Clients.ActiveDirectory/Platforms/uap/PlatformInformation.cs
@@ -47,106 +47,6 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory.Internal.Platform
             return "PCL.UAP";
         }
 
-        public override string GetEnvironmentVariable(string variable)
-        {
-            ApplicationDataContainer localSettings = ApplicationData.Current.LocalSettings;
-            return localSettings.Values.ContainsKey(variable) ? localSettings.Values[variable].ToString() : null;
-        }
-
-        /// <summary>
-        /// Get a principal name that can be used for WIA
-        /// </summary>
-        /// <remarks>
-        /// Win10 allows several identities to be logged in at once; 
-        /// select the first principal name that can be used
-        /// </remarks>
-        /// <returns></returns>
-        public override async Task<string> GetUserPrincipalNameAsync()
-        {
-            IReadOnlyList<User> users = await User.FindAllAsync();
-            if (users == null || !users.Any())
-            {
-                throw new AdalException(
-                 AdalErrorEx.CannotAccessUserInformationOrUserNotDomainJoined,
-                 CoreErrorMessages.UapCannotFindDomainUser);
-            }
-
-            var getUserDetailTasks = users.Select(async u =>
-            {
-                object domainObj = await u.GetPropertyAsync(KnownUserProperties.DomainName);
-                string domainString = domainObj?.ToString();
-
-                object principalObject = await u.GetPropertyAsync(KnownUserProperties.PrincipalName);
-                string principalNameString = principalObject?.ToString();
-
-                return new { Domain = domainString, PrincipalName = principalNameString };
-            }).ToList();
-
-            var userDetails = await Task.WhenAll(getUserDetailTasks).ConfigureAwait(false);
-
-            // try to get a user that has both domain name and upn
-            var userDetailWithDomainAndPn = userDetails.FirstOrDefault(
-                d => !String.IsNullOrWhiteSpace(d.Domain) &&
-                !String.IsNullOrWhiteSpace(d.PrincipalName));
-
-            if (userDetailWithDomainAndPn != null)
-            {
-                return userDetailWithDomainAndPn.PrincipalName;
-            }
-
-            // try to get a user that at least has upn
-            var userDetailWithPn = userDetails.FirstOrDefault(
-              d => !String.IsNullOrWhiteSpace(d.PrincipalName));
-
-            if (userDetailWithPn != null)
-            {
-                return userDetailWithPn.PrincipalName;
-            }
-
-            // user has domain name, but no upn -> missing Enterprise Auth capability
-            if (userDetails.Any(d => !String.IsNullOrWhiteSpace(d.Domain)))
-            {
-                throw new AdalException(
-                 CoreErrorCodes.UapCannotFindUpn,
-                 CoreErrorMessages.UapCannotFindUpn);
-            }
-
-            // no domain, no upn -> missing User Info capability
-            throw new AdalException(
-             CoreErrorMessages.UapCannotFindDomainUser,
-             CoreErrorMessages.UapCannotFindDomainUser);
-
-        }
-
-        public override string GetProcessorArchitecture()
-        {
-            return WindowsNativeMethods.GetProcessorArchitecture();
-        }
-
-        public override string GetOperatingSystem()
-        {
-            // In WinRT, there is no way to reliably get OS version. All can be done reliably is to check 
-            // for existence of specific features which does not help in this case, so we do not emit OS in WinRT.
-            return null;
-        }
-
-        public override string GetDeviceModel()
-        {
-            var deviceInformation = new Windows.Security.ExchangeActiveSyncProvisioning.EasClientDeviceInformation();
-            return deviceInformation.SystemProductName;
-        }
-
-        public override async Task<bool> IsUserLocalAsync(RequestContext requestContext)
-        {
-            IReadOnlyList<User> users = await User.FindAllAsync();
-            return users.Any(u => u.Type == UserType.LocalUser || u.Type == UserType.LocalGuest);
-        }
-
-        public override bool IsDomainJoined()
-        {
-            return NetworkInformation.GetHostNames().Any(entry => entry.Type == HostNameType.DomainName);
-        }
-
         public override void AddPromptBehaviorQueryParameter(IPlatformParameters parameters, DictionaryRequestParameters authorizationRequestParameters)
         {
             PlatformParameters authorizationParameters = (parameters as PlatformParameters);
