+++ core/src/CorePlatformInformationBase.cs
@@ -36,24 +36,8 @@ namespace Microsoft.Identity.Core
 
         public const string DefaultRedirectUri = "urn:ietf:wg:oauth:2.0:oob";
         public abstract string GetProductName();
-        public virtual Task<string> GetUserPrincipalNameAsync()
-        {
-            return Task.FromResult<string>(null);
-        }
-
-        public abstract string GetEnvironmentVariable(string variable);
-        public abstract string GetProcessorArchitecture();
-        public abstract string GetOperatingSystem();
-        public abstract string GetDeviceModel();
 
         public abstract string GetAssemblyFileVersionAttribute();
-
-        public abstract Task<bool> IsUserLocalAsync(RequestContext requestContext);
-
-        public virtual bool IsDomainJoined()
-        {
-            return false;
-        }
 
         public virtual void ValidateRedirectUri(Uri redirectUri, RequestContext requestContext)
         {
