+++ core/src/Exceptions/CoreErrorCodes.cs
@@ -34,6 +34,8 @@ namespace Microsoft.Identity.Core
     /// </summary>
     internal class CoreErrorCodes
     {
+
+
         public const string JsonParseError = "json_parse_failed";
         public const string RequestTimeout = "request_timeout";
         public const string ServiceNotAvailable = "service_not_available";
@@ -69,10 +71,12 @@ namespace Microsoft.Identity.Core
         public const string WsTrustEndpointNotFoundInMetadataDocument = "wstrust_endpoint_not_found";
         public const string ParsingWsMetadataExchangeFailed = "parsing_ws_metadata_exchange_failed";
         public const string UnknownUserType = "unknown_user_type";
+        public const string CannotAccessUserInformationOrUserNotDomainJoined = "user_information_access_failed";
 
         public const string UapCannotFindDomainUser = "user_information_access_failed";
         public const string UapCannotFindUpn = "uap_cannot_find_upn";
 
+        public const string GetUserNameFailed = "get_user_name_failed";
 
 
 #if ANDROID
