+++ core/src/Exceptions/CoreErrorMessages.cs
@@ -88,5 +88,8 @@ namespace Microsoft.Identity.Core
             "Internal error - trying to remove an ADAL user with an empty username. Possible cache corruption. See https://aka.ms/adal_token_cache_serialization";
         public const string InternalErrorCacheEmptyIdentifier = 
             "Internal error - trying to remove an ADAL user with an empty identifier. Possible cache corruption. See https://aka.ms/adal_token_cache_serialization";
+
+        public const string GetUserNameFailed = "Failed to get user name from the operating system.";
+
     }
 }
