+++ core/src/Platforms/netcore/PlatformProxy.cs
@@ -25,14 +25,50 @@
 //
 //------------------------------------------------------------------------------
 
-using System;
 using System.Threading.Tasks;
 
-namespace Microsoft.IdentityModel.Clients.ActiveDirectory.Internal.Platform
+namespace Microsoft.Identity.Core
 {
-    internal class PlatformInformation : PlatformInformationBase
+    /// <summary>
+    /// Platform / OS specific logic.  No library (ADAL / MSAL) specific code should go in here. 
+    /// </summary>
+    internal class PlatformProxy : IPlatformProxy
     {
-        public override string GetProductName()
+        /// <summary>
+        /// Get the user logged in 
+        /// </summary>
+        public async Task<string> GetUserPrincipalNameAsync()
+        {
+            return await Task.Factory.StartNew(() => string.Empty).ConfigureAwait(false);
+
+        }
+        public async Task<bool> IsUserLocalAsync(RequestContext requestContext)
+        {
+            return await Task.Factory.StartNew(() => false).ConfigureAwait(false);
+        }
+
+        public bool IsDomainJoined()
+        {
+            return false;
+        }
+
+        public string GetEnvironmentVariable(string variable)
+        {
+            string value = System.Environment.GetEnvironmentVariable(variable);
+            return !string.IsNullOrWhiteSpace(value) ? value : null;
+        }
+
+        public string GetProcessorArchitecture()
+        {
+            return null;
+        }
+
+        public string GetOperatingSystem()
+        {
+            return System.Runtime.InteropServices.RuntimeInformation.OSDescription;
+        }
+
+        public string GetDeviceModel()
         {
             return null;
         }
