+++ core/src/Platforms/netstandard1.1/PlatformProxy.cs
@@ -25,16 +25,54 @@
 //
 //------------------------------------------------------------------------------
 
-using System;
 using System.Threading.Tasks;
 
-namespace Microsoft.IdentityModel.Clients.ActiveDirectory.Internal.Platform
+namespace Microsoft.Identity.Core
 {
-    internal class PlatformInformation : PlatformInformationBase
+    /// <summary>
+    /// Platform / OS specific logic.  No library (ADAL / MSAL) specific code should go in here. 
+    /// </summary>
+    internal class PlatformProxy : IPlatformProxy
     {
-        public override string GetProductName()
+        /// <summary>
+        /// Get the user logged in 
+        /// </summary>
+        public  async Task<string> GetUserPrincipalNameAsync()
+        {
+            return await Task.Factory.StartNew(() => string.Empty).ConfigureAwait(false);
+
+        }
+
+        public  async Task<bool> IsUserLocalAsync(RequestContext requestContext)
+        {
+            return await Task.Factory.StartNew(() => false).ConfigureAwait(false);
+        }
+
+        public  bool IsDomainJoined()
+        {
+            return false;
+        }
+
+        public  string GetEnvironmentVariable(string variable)
         {
             return null;
         }
+
+        public  string GetProcessorArchitecture()
+        {
+            return null;
+        }
+
+        public  string GetOperatingSystem()
+        {
+            return null;
+        }
+
+        public  string GetDeviceModel()
+        {
+            return null;
+        }
+
+
     }
 }
