+++ msal/src/Microsoft.Identity.Client/Platforms/net45/PlatformInformation.cs
@@ -25,126 +25,17 @@
 //
 //------------------------------------------------------------------------------
 
-using System;
-using System.ComponentModel;
-using System.Runtime.InteropServices;
-using System.Security.Principal;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.Identity.Client.Internal;
-using Microsoft.Identity.Core;
-using Microsoft.Identity.Core.Platforms;
 
 namespace Microsoft.Identity.Client
 {
     internal class PlatformInformation : PlatformInformationBase
     {
-	internal static bool IsWindows {
-            get {
-                switch (Environment.OSVersion.Platform) {
-                    case PlatformID.Win32S:
-                    case PlatformID.Win32Windows:
-                    case PlatformID.Win32NT:
-                    case PlatformID.WinCE:
-                        return true;
-                    default:
-                        return false;
-		}
-            }
-        }
-
+      
         public override string GetProductName()
         {
             return "MSAL.Desktop";
         }
 
-        public override async Task<string> GetUserPrincipalNameAsync()
-        {
-            return await Task.Factory.StartNew(() =>
-            {
-                const int NameUserPrincipal = 8;
-                uint userNameSize = 0;
-                WindowsNativeMethods.GetUserNameEx(NameUserPrincipal, null, ref userNameSize);
-                if (userNameSize == 0)
-                {
-                    throw new MsalClientException(MsalError.GetUserNameFailed, MsalError.GetUserNameFailed, new Win32Exception(Marshal.GetLastWin32Error()));
-                }
-
-                StringBuilder sb = new StringBuilder((int)userNameSize);
-                if (!WindowsNativeMethods.GetUserNameEx(NameUserPrincipal, sb, ref userNameSize))
-                {
-                    throw new MsalClientException(MsalError.GetUserNameFailed, MsalError.GetUserNameFailed, new Win32Exception(Marshal.GetLastWin32Error()));
-                }
-
-                return sb.ToString();
-            }).ConfigureAwait(false);
-        }
-
-        public override string GetEnvironmentVariable(string variable)
-        {
-            string value = Environment.GetEnvironmentVariable(variable);
-            return !string.IsNullOrWhiteSpace(value) ? value : null;
-        }
-
-        public override string GetProcessorArchitecture()
-        {
-            return IsWindows ? WindowsNativeMethods.GetProcessorArchitecture() : null;
-        }
-
-        public override string GetOperatingSystem()
-        {
-            return Environment.OSVersion.ToString();
-        }
-
-        public override string GetDeviceModel()
-        {
-            // Since MSAL .NET may be used on servers, for security reasons, we do not emit device type.
-            return null;
-        }
-
-        public override async Task<bool> IsUserLocalAsync(RequestContext requestContext)
-        {
-            return await Task.Factory.StartNew(() =>
-            {
-                WindowsIdentity current = WindowsIdentity.GetCurrent();
-                if (current != null)
-                {
-                    string prefix = WindowsIdentity.GetCurrent().Name.Split('\\')[0].ToUpperInvariant();
-                    return prefix.Equals(Environment.MachineName.ToUpperInvariant(), StringComparison.OrdinalIgnoreCase);
-                }
-
-                return false;
-            }).ConfigureAwait(false);
-        }
-
-        public override bool IsDomainJoined()
-        {
-            if (!IsWindows)
-            {
-                return false;
-            }
-
-            bool returnValue = false;
-            try
-            {
-                WindowsNativeMethods.NetJoinStatus status;
-                IntPtr pDomain;
-                int result = WindowsNativeMethods.NetGetJoinInformation(null, out pDomain, out status);
-                if (pDomain != IntPtr.Zero)
-                {
-                    WindowsNativeMethods.NetApiBufferFree(pDomain);
-                }
-
-                returnValue = result == WindowsNativeMethods.ErrorSuccess &&
-                              status == WindowsNativeMethods.NetJoinStatus.NetSetupDomainName;
-            }
-            catch (Exception ex)
-            {
-                CoreLoggerBase.Default.WarningPii(ex);
-                // ignore the exception as the result is already set to false;
-            }
-
-            return returnValue;
-        }
     }
 }
